Added redirection to author of post on shared

+++ src/lib/server/services/post.service.ts
@@ -10,31 +10,61 @@ export class PostService {
 
   async getAllPosts(options: GetAllPostsOptions = {}): Promise<Post[]> {
     const userId = await getCurrentUserId();
+
+    if (userId == null) {
+      throw new Error("user not found");
+    }
+
     return this.postRepository.getAll(userId, options);
   }
 
   async getPostById(id: string): Promise<PostWithUser | null> {
     const userId = await getCurrentUserId();
+
+    if (userId == null) {
+      throw new Error("user not found");
+    }
+
     return this.postRepository.getById(userId, id);
   }
 
   async getPostBySlug(slug: string): Promise<PostWithUser | null> {
     const userId = await getCurrentUserId();
+
+    if (userId == null) {
+      throw new Error("user not found");
+    }
+
     return this.postRepository.getBySlug(userId, slug);
   }
 
   async createPost(post: CreatePost): Promise<Post> {
     const userId = await getCurrentUserId();
+
+    if (userId == null) {
+      throw new Error("user not found");
+    }
+
     return this.postRepository.create(userId, post);
   }
 
   async updatePost(post: UpdatePost): Promise<Post | null> {
     const userId = await getCurrentUserId();
+
+    if (userId == null) {
+      throw new Error("user not found");
+    }
+
     return this.postRepository.update(userId, post);
   }
 
   async deletePost(id: string): Promise<Post | null> {
     const userId = await getCurrentUserId();
+
+    if (userId == null) {
+      throw new Error("user not found");
+    }
+
     return this.postRepository.delete(userId, id);
   }
 }
